+++ examples/03_connectivity/plot_data_driven_parcellations.py
@@ -2,8 +2,8 @@
 Clustering methods to learn a brain parcellation from fMRI
 ==========================================================
 
-We use spatially-constrained Ward-clustering, KMeans, and Recursive Neighbor
-Agglomeration (ReNA) to create a set of parcels.
+We use spatially-constrained Ward-clustering, KMeans, Hierarchical KMeans
+and Recursive Neighbor Agglomeration (ReNA) to create a set of parcels.
 
 In a high dimensional regime, these methods can be interesting
 to create a 'compressed' representation of the data, replacing the data
@@ -11,13 +11,14 @@ in the fMRI images by mean signals on the parcellation, which can
 subsequently be used for statistical analysis or machine learning.
 
 Also, these methods can be used to learn functional connectomes
-and subsequently for classification tasks.
+and subsequently for classification tasks or to analyze data at a local
+level.
 
 References
 ----------
 
 Which clustering method to use, an empirical comparison can be found in this
-paper
+paper:
 
     * Bertrand Thirion, Gael Varoquaux, Elvis Dohmatob, Jean-Baptiste Poline.
       `Which fMRI clustering gives good brain parcellations ?
@@ -25,7 +26,7 @@ paper
       2014.
 
 This parcellation may be useful in a supervised learning, see for
-instance
+instance:
 
     * Vincent Michel, Alexandre Gramfort, Gael Varoquaux, Evelyn Eger,
       Christine Keribin, Bertrand Thirion. `A supervised clustering approach
@@ -43,6 +44,14 @@ in the documentation section :ref:`parcellating_brain`.
 #
 # We download one subject of the movie watching dataset from Internet
 
+from matplotlib import patches, ticker
+import matplotlib.pyplot as plt
+from nilearn.image import get_data
+import numpy as np
+from nilearn.image import mean_img, index_img
+from nilearn import plotting
+import time
+from nilearn.regions import Parcellations
 from nilearn import datasets
 dataset = datasets.fetch_development_fmri(n_subjects=1)
 
@@ -58,11 +67,9 @@ print('First subject functional nifti image (4D) is at: %s' %
 # Transforming list of images to data matrix and build brain parcellations,
 # all can be done at once using `Parcellations` object.
 
-from nilearn.regions import Parcellations
 
 # Computing ward for the first time, will be long... This can be seen by
 # measuring using time
-import time
 start = time.time()
 
 # Agglomerative Clustering: ward
@@ -102,8 +109,6 @@ ward_labels_img = ward.labels_img_
 # with the following code:
 ward_labels_img.to_filename('ward_parcellation.nii.gz')
 
-from nilearn import plotting
-from nilearn.image import mean_img, index_img
 
 first_plot = plotting.plot_roi(ward_labels_img, title="Ward parcellation",
                                display_mode='xz')
@@ -120,8 +125,6 @@ cut_coords = first_plot.cut_coords
 # clustering by averaging the signal on each parcel.
 
 # Grab number of voxels from attribute mask image (mask_img_).
-import numpy as np
-from nilearn.image import get_data
 original_voxels = np.sum(get_data(ward.mask_img_))
 
 # Compute mean over time on the functional image to use the mean
@@ -183,15 +186,109 @@ print("KMeans clusters: %.2fs" % (time.time() - start))
 # Grab parcellations of brain image stored in attribute `labels_img_`
 kmeans_labels_img = kmeans.labels_img_
 
-plotting.plot_roi(kmeans_labels_img, mean_func_img,
-                  title="KMeans parcellation",
-                  display_mode='xz')
+display = plotting.plot_roi(kmeans_labels_img, mean_func_img,
+                            title="KMeans parcellation",
+                            display_mode='xz')
 
 # kmeans_labels_img is a Nifti1Image object, it can be saved to file with
 # the following code:
 kmeans_labels_img.to_filename('kmeans_parcellation.nii.gz')
 
-##################################################################
+#########################################################################
+# Brain parcellations with Hierarchical KMeans Clustering
+# -------------------------------------------------------
+#
+# As the number of images from which we try to cluster grows,
+# voxels display more and more specific activity patterns causing
+# KMeans clusters to be very unbalanced with a few big clusters and
+# many voxels left as singletons. Hierarchical Kmeans algorithm is
+# tailored to enforce more balanced clusterings. To do this,
+# Hierarchical Kmeans does a first Kmeans clustering in square root of
+# n_parcels. In a second step, it clusters voxels inside each
+# of these parcels in m pieces with m adapted to the size of
+# the cluster in order to have n balanced clusters in the end.
+#
+# This object uses method='hierarchical_kmeans' for Hierarchical KMeans
+# clustering and 10mm smoothing and standardization to compare
+# with the previous method.
+start = time.time()
+hkmeans = Parcellations(method='hierarchical_kmeans', n_parcels=50,
+                        standardize=True, smoothing_fwhm=10,
+                        memory='nilearn_cache', memory_level=1,
+                        verbose=1)
+# Call fit on functional dataset: single subject (less samples)
+hkmeans.fit(dataset.func)
+
+###########################################################################
+# Visualize: Brain parcellations (Hierarchical KMeans)
+# ....................................................
+#
+# Grab parcellations of brain image stored in attribute `labels_img_`
+hkmeans_labels_img = hkmeans.labels_img_
+
+plotting.plot_roi(hkmeans_labels_img, mean_func_img,
+                  title="Hierarchical KMeans parcellation",
+                  display_mode='xz', cut_coords=display.cut_coords)
+
+# kmeans_labels_img is a :class:`nibabel.nifti1.Nifti1Image` object, it can be
+# saved to file with the following code:
+hkmeans_labels_img.to_filename('hierarchical_kmeans_parcellation.nii.gz')
+
+###########################################################################
+# Compare Hierarchical Kmeans clusters with those from Kmeans
+# ...........................................................
+# To compare those, we'll first count how many voxels are contained in
+# each of the 50 clusters for both algorithms and compare those sizes
+# distribution. Hierarchical KMeans should give clusters closer to
+# average (600 here) than KMeans.
+#
+# First count how many voxels have each label (except 0 which is the
+# background).
+
+kmeans_labels, kmeans_counts = np.unique(
+    get_data(kmeans_labels_img), return_counts=True)
+
+_, hkmeans_counts = np.unique(
+    get_data(hkmeans_labels_img), return_counts=True)
+
+voxel_ratio = np.round(np.sum(kmeans_counts[1:]) / 50)
+
+# If all voxels not in background were balanced between clusters ...
+
+print("... each cluster should contain {} voxels".format(voxel_ratio))
+
+###########################################################################
+# Let's plot clusters sizes distributions for both algorithms
+#
+# You can just skip the plotting code, the important part is the figure
+
+
+bins = np.concatenate([np.linspace(0, 500, 11), np.linspace(
+    600, 2000, 15), np.linspace(3000, 10000, 8)])
+fig, axes = plt.subplots(nrows=2, sharex=True, gridspec_kw={
+                         'height_ratios': [4, 1]})
+plt.semilogx()
+axes[0].hist(kmeans_counts[1:], bins, color="blue")
+axes[1].hist(hkmeans_counts[1:], bins, color="green")
+axes[0].set_ylim(0, 16)
+axes[1].set_ylim(4, 0)
+axes[1].xaxis.set_major_formatter(ticker.ScalarFormatter())
+axes[1].yaxis.set_label_coords(-.08, 2)
+fig.subplots_adjust(hspace=0)
+plt.xlabel("Number of voxels (log)", fontsize=12)
+plt.ylabel("Number of clusters", fontsize=12)
+handles = [patches.Rectangle((0, 0), 1, 1, color=c, ec="k")
+           for c in ["blue", "green"]]
+labels = ["Kmeans", "Hierarchical Kmeans"]
+fig.legend(handles, labels, loc=(.5, .8))
+###########################################################################
+# As we can see, half of the 50 KMeans clusters contain less than
+# 100 voxels whereas three contain several thousands voxels
+# Hierarchical KMeans yield better balanced clusters, with a significant
+# proportion of them containing hundreds to thousands of voxels.
+#
+
+###########################################################################
 # Brain parcellations with ReNA Clustering
 # ----------------------------------------
 #
@@ -218,7 +315,7 @@ rena = Parcellations(method='rena', n_parcels=5000, standardize=False,
 rena.fit_transform(dataset.func)
 print("ReNA 5000 clusters: %.2fs" % (time.time() - start))
 
-##################################################################
+###########################################################################
 # Visualize: Brain parcellations (ReNA)
 # .....................................
 #
@@ -233,7 +330,7 @@ rena_labels_img.to_filename('rena_parcellation.nii.gz')
 plotting.plot_roi(ward_labels_img, title="ReNA parcellation",
                   display_mode='xz', cut_coords=cut_coords)
 
-##################################################################
+###########################################################################
 # Compressed representation of ReNA clustering
 # ............................................
 #
@@ -264,7 +361,7 @@ plotting.plot_epi(index_img(compressed_img_rena, 0), cut_coords=cut_coords,
                   title='ReNA compressed representation (5000 parcels)',
                   vmin=vmin, vmax=vmax, display_mode='xz')
 
-####################################################################
+###########################################################################
 # Even if the compressed signal is relatively close
 # to the original signal, we can notice that Ward Clustering
 # gives a slightly more accurate compressed representation.